requirements.txt

Django==2.2.3
gunicorn==20.0.4
idna==2.9
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
whitenoise==5.0.1
python-decouple==3.3      
python-decouple==3.3
psycopg2-binary==2.8.3
dj-database-url==0.5.0    
# django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0    
django-cors-headers==3.2.1 